+++ audioengine.h
@@ -26,7 +26,7 @@ public:
 	void set_output_volume(qreal);
 	void set_input_volume(qreal);
 	bool frame_available() { return (m_audioinq.size() >= 320) ? true : false; }
-	uint16_t read(int16_t *, size_t);
+	uint16_t read(int16_t *, int);
 signals:
 
 private:
